python_backend/requirements.txt
qiskit>=1.0.0
qiskit-ibm-runtime>=0.22.0
fastapi
uvicorn
python-dotenv